fix font and ressources
