+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="../../../../../layout/compose-model-1650475049600.xml" value="0.13851351351351351" />
         <entry key="../../../../../layout/compose-model-1650475104697.xml" value="1.2244389027431422" />
         <entry key="../../../../../layout/compose-model-1650475116766.xml" value="2.0" />
+        <entry key="../../../../../layout/compose-model-1650556040550.xml" value="0.1089527027027027" />
         <entry key="../../../../layout/compose-model-1650116345079.xml" value="1.0" />
         <entry key="../../../../layout/compose-model-1650123853863.xml" value="1.4311377245508983" />
         <entry key="../../../../layout/compose-model-1650128533389.xml" value="1.0" />
