+++ app/src/androidTest/java/io/webthings/app/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package io.webthings.webthings
+package io.webthings.app
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
