+++ app/src/main/java/io/webthings/app/ui/NavRoutes.kt
@@ -1,23 +1,30 @@
-package io.webthings.webthings.ui
+package io.webthings.app.ui
 
 
+import android.content.Context
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Face
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.ui.graphics.vector.ImageVector
+import io.webthings.app.R
+import android.content.res.Resources
+
+private val _context = Resources.getSystem()
 
 sealed class NavRoutes(val route:String, val title:String){
+
     object HomeScreen : NavRoutes("home",
-        "home")
+        _context.getString(R.string.home))
 
     sealed class DrawerItems(
         route: String,
         title: String,
         val icon: ImageVector): NavRoutes(route, title){
         object DashboardScreen: DrawerItems("Dashboard",
-            "Main Dashboard", Icons.Default.Face)
+            _context.getString(R.string.main_dashboard),
+            Icons.Default.Face)
             object GwSettingsScreen: DrawerItems("GwSettings",
-                "Settings",
+                _context.getString(R.string.Settings),
                 Icons.Default.Settings)
         }
