+++ app/src/main/java/io/webthings/app/ui/menus/Drawer.kt
@@ -1,4 +1,4 @@
-package io.webthings.webthings.ui.menus
+package io.webthings.app.ui.menus
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -9,10 +9,12 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import io.webthings.webthings.ui.drawerMenuItems
-import io.webthings.webthings.ui.theme.*
+import io.webthings.app.ui.drawerMenuItems
+import io.webthings.app.ui.theme.*
+import io.webthings.app.R
 
 @Composable
 fun Drawer(
@@ -24,7 +26,7 @@ fun Drawer(
                         .padding(top = 10.dp)
         ) {
                 Text(
-                        "Webthings",
+                        stringResource(R.string.app_name),
                         style = HeaderStyle,
                         modifier = Modifier
                                 .padding(bottom = 10.dp)
@@ -48,7 +50,8 @@ fun Drawer(
                                                 modifier = Modifier
                                                         .clickable {
                                                                 onDestinationClicked(
-                                                                        screen.route)
+                                                                        screen.route
+                                                                )
                                                         }
                                                         .padding(start = 24.dp)
                                         )
