+++ app/src/main/java/io/webthings/app/ui/screens/DashboardScreen.kt
@@ -1,12 +1,12 @@
-package io.webthings.webthings.ui.screens
+package io.webthings.app.ui.screens
 
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import io.webthings.webthings.DEVICES_PREVIEW
-import io.webthings.webthings.ui.NavRoutes
-import io.webthings.webthings.ui.theme.WebthingsTheme
-import io.webthings.webthings.utils.MainViewModel
+import io.webthings.app.DEVICES_PREVIEW
+import io.webthings.app.ui.NavRoutes
+import io.webthings.app.ui.theme.WebthingsTheme
+import io.webthings.app.utils.MainViewModel
 
 @Composable
 fun DashboardScreen(viewModel: MainViewModel) {
