+++ app/src/main/java/io/webthings/app/ui/screens/GwSettingsScreen.kt
@@ -1,10 +1,10 @@
-package io.webthings.webthings.ui.screens
+package io.webthings.app.ui.screens
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import io.webthings.webthings.ui.NavRoutes
-import io.webthings.webthings.utils.MainViewModel
+import io.webthings.app.ui.NavRoutes
+import io.webthings.app.utils.MainViewModel
 
 @Composable
 fun GwSettingsScreen(viewModel: MainViewModel){
