+++ app/src/main/java/io/webthings/app/ui/screens/HomeScreen.kt
@@ -1,14 +1,11 @@
-package io.webthings.webthings.ui.screens
+package io.webthings.app.ui.screens
 
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import io.webthings.webthings.DEVICES_PREVIEW
-import io.webthings.webthings.ui.NavRoutes
-import io.webthings.webthings.ui.theme.WebthingsTheme
+import io.webthings.app.DEVICES_PREVIEW
+import io.webthings.app.ui.theme.WebthingsTheme
 
 @Composable
 fun HomeScreen(navController: NavController){
