+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Webthings</string>
+    <string name="app_name">WebThings</string>
 </resources>
