+++ app/src/main/res/values/strings_drawer.xml
@@ -2,4 +2,5 @@
 <resources>
     <string name="home">Home</string>
     <string name="main_dashboard">Main Dashboard</string>
+    <string name="Settings">Settings</string>
 </resources>
